RELEASE 1.0

+++ my_health_service/src/main/java/com/lwy/service/MenuServiceImpl.java
@@ -1,7 +1,11 @@
 package com.lwy.service;
 
 import com.alibaba.dubbo.config.annotation.Service;
+import com.github.pagehelper.Page;
+import com.github.pagehelper.PageHelper;
 import com.lwy.dao.MenuDao;
+import com.lwy.entity.PageResult;
+import com.lwy.entity.QueryPageBean;
 import com.lwy.exception.HealthException;
 import com.lwy.pojo.Menu;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +29,7 @@ public class MenuServiceImpl  implements MenuService{
     }
 
     @Override
-    public void add(Menu menu) {
+    public void add(Menu menu) throws HealthException{
         Integer parentMenuId = menu.getParentMenuId();
         if (parentMenuId!=null){
             Menu parentMenu=menuDao.findById(parentMenuId);
@@ -37,12 +41,16 @@ public class MenuServiceImpl  implements MenuService{
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Integer id)throws HealthException {
+        long count =menuDao.findCountByMenuId(id);
+        if (count>0){
+            throw new HealthException("菜单已被引用");
+        }
         menuDao.deleteById(id);
     }
 
     @Override
-    public void edit(Menu menu) {
+    public void edit(Menu menu)throws HealthException {
         Integer parentMenuId = menu.getParentMenuId();
         if (parentMenuId!=null){
             Menu parentMenu=menuDao.findById(parentMenuId);
@@ -53,4 +61,20 @@ public class MenuServiceImpl  implements MenuService{
         menuDao.edit(menu);
     }
 
+    @Override
+    public PageResult findPage(QueryPageBean queryPageBean) {
+        String queryString = queryPageBean.getQueryString();
+        if (queryString!=null){
+            queryString="%"+queryString+"%";
+            queryPageBean.setCurrentPage(1);
+        }
+        PageHelper.startPage(queryPageBean.getCurrentPage(), queryPageBean.getPageSize());
+        Page<Menu> page= menuDao.findByCondition(queryString);
+        return new PageResult(page.getTotal(),page.getResult());
+    }
+
+    @Override
+    public Menu findById(Integer id) {
+        return menuDao.findById(id);
+    }
 }
